ApartmentParser is moved to separate module
